refactor per discussion
