+++ neural_compressor/torch/algorithms/habana_fp8/__init__.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2024 Intel Corporation
+# Copyright (c) 2023 Intel Corporation
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -12,7 +12,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .quantization_impl import (
-    rtn_entry,
-    gptq_entry,
-)
+from .fp8_quant import quantize_dynamic, quantize, white_list
