+++ neural_compressor/torch/quantization/quantize.py
@@ -66,44 +66,3 @@ def quantize(
             logger.info(f"Start to apply {algo_name} on the model.")
             q_model = algo_func(q_model, configs_mapping, run_fn=run_fn, run_args=run_args)
     return q_model
-
-
-def quantize_dynamic(
-    model: torch.nn.Module,
-    quant_config: BaseConfig,
-    run_fn: Callable = None,
-    run_args: Any = None,
-    inplace: bool = True,
-) -> torch.nn.Module:
-    """The main entry to quantize model with dynamic mode.
-
-    Args:
-        model: a float model to be quantized.
-        quant_config: a quantization configuration.
-        run_fn: a calibration function for calibrating the model. Defaults to None.
-        run_args: positional arguments for `run_fn`. Defaults to None.
-        inplace: carry out model transformations in-place, the original module is mutated. Default is True.
-
-    Returns:
-        The quantized model.
-    """
-    q_model = model if inplace else copy.deepcopy(model)
-    registered_configs = config_registry.get_cls_configs()
-    if isinstance(quant_config, dict):
-        quant_config = ComposableConfig.from_dict(quant_config, config_registry=registered_configs[FRAMEWORK_NAME])
-        logger.info(f"Parsed a config dict to construct the quantization config: {quant_config}.")
-    else:
-        assert isinstance(
-            quant_config, BaseConfig
-        ), "Please pass a dict or config instance as the quantization configuration."
-    logger.info(f"Quantize model with config: \n {quant_config.to_json_string()} \n")
-    # select quantization algo according to config
-
-    model_info = quant_config.get_model_info(model=q_model)
-    configs_mapping = quant_config.to_config_mapping(model_info=model_info)
-    logger.debug(configs_mapping)
-    for algo_name, algo_func in algos_mapping.items():
-        if need_apply(configs_mapping, algo_name):
-            logger.info(f"Start to apply {algo_name} on the model.")
-            q_model = algo_func(q_model, configs_mapping, run_fn=run_fn, run_args=run_args)
-    return q_model
